fix error on delete route

+++ src/app/api/deleteParticipation/route.ts
@@ -3,6 +3,9 @@ import prisma from "../../../../prisma/prisma";
 
 export async function DELETE(request: NextRequest) {
   const user = request.nextUrl.searchParams.get("user");
+  if (user === null) {
+    return NextResponse.json({ message: "user is null" });
+  }
   const now = new Date();
   const currentHour = now.getHours();
   const deleteResult = await prisma.user.deleteMany({
